Added v1.2 Files
These are the direct files I created to run the new processes for Gemini. I can't upload a large file like I did before, but these are the changed files. This new update includes a Discord but that has a continued conversation feature. It makes a file of the conversation history, then rereads it everytime the file is ran the the Gemini is turned on. You have to remind it of the previous conversation but it does work. It also includes the Image scan feature and Document scan feature. This is in the CLI version only unfortunately, but it will be included in the discord bot in v1.3. It also contains the beginning and very heavily in development file that you can use to train a gemini model. This is more of an incomplete framework, but you can continue it or modify it. Also, in the CLI version, there will be a history.txt file that will contain the entire conversation history of gemini in CLI. This is done for testing purposes with continued conversation and can be removed. I know me code is messy but I am trying. Thank you